Add the InvoiceInteractor
This class is responsible to abstract all the business rules to make some actions. e.g: To create a invoice, we need to do a raffle of the Payers first.
So the createInvoiceUseCase covers not only the proper creation of the invoice, but also the raffle of the customer Payers.